POST Data is important!
Before redirecting users to the login page (if required), pages that
require login now check if there was any post data before sending the
user to the login page. I know it's not ideal, but this issue only
occurs when the app restarts before a user can save their data. The
rarity of this has influenced my decision to not add even more
validators to every post route